Handle cases where pulling fails silently.
Sometimes, on various versions of Docker, the pull fails silently. The
actual image create does seemingly nothing and throws no exception, and
our subsequent attempt to check the image is locally returns `nil`. Up
until now, we never guarded against this `nil` properly. Now we do so
within the `pull` so that no matter how it failed, if after `pull`ing we
couldn't find the image, we raise an exception.